Added train_advanced_single.py, which takes the functionality of the train_advanced.py but uses a single environment over a vectorized environement. Datetime now used to better name logs.
Slight changes may need to be made depending on your algorithm, such as the inclusion of noise
(which is currently commeneted out) and the changing of variable names. Note that the commented lines regarding videoing evaluations
are not yet modified to fit the single environment and if you uncomment them you will likely get an error.